requirements.txt
astroid==1.4.5
colorama==0.3.7
Django==1.9.6
factory-boy==2.7.0
fake-factory==0.5.7
lazy-object-proxy==1.2.2
py==1.4.31
pylint==1.5.5
pytest==2.9.1
python-dateutil==2.5.3
six==1.10.0
wrapt==1.10.8
